Full functional test

+++ GameApi/Controllers/GamesController.cs
@@ -49,15 +49,28 @@ namespace GameApi.Controllers
         /// <response code="404">Creator not found in player list</response>
         [HttpPost]
         [ProducesResponseType(typeof(GameDto), StatusCodes.Status201Created)]
+        [ProducesResponseType(typeof(GameDto), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(GameDto), StatusCodes.Status404NotFound)]
         public ActionResult<GameDto?> CreateGame([FromBody] GameDto game)
         {
             if(game.Creator is null || game.Title is null)
             {
                 return BadRequest("Title or creator id missing");
             }
-            var g = _manager.NewGame(game.Title, game.Creator.Value);
+            try
+            {
+                var g = _manager.NewGame(game.Title, game.Creator.Value);
 
-            return Created($"/games/{g.Id}", GameDto.FromGame(g));
+                return Created($"/games/{g.Id}", GameDto.FromGame(g));
+            }
+            catch(ArgumentException)
+            {
+                return NotFound("No player found with this private id");
+            }
+            catch(InvalidOperationException)
+            {
+                return Conflict("Creator already in a game");
+            }
         }
 
         /// <summary>
@@ -67,13 +80,20 @@ namespace GameApi.Controllers
         /// <param name="creatorPrivateId">Private Id of the game creator</param>
         /// <returns>No content</returns>
         /// <response code="204">Game successfully deleted</response>
-        /// <response code="400">The given id is not the creator's private id</response>
+        /// <response code="400">Creator Private Id not set</response>
+        /// <response code="401">The given id is not the creator's private id</response>
         /// <response code="404">Game not found</response>
         [HttpDelete("{gameId:guid}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public ActionResult<GameDto?> DeleteGame(Guid gameId, Guid creatorPrivateId)
         {
+            if(creatorPrivateId == Guid.Empty)
+            {
+                return BadRequest("Missing the creator private Id parameter");
+            }
             try
             {
                 _manager.DeleteGame(gameId, creatorPrivateId);
@@ -84,7 +104,7 @@ namespace GameApi.Controllers
                 return ex.ParamName switch
                 {
                     "gameId" => NotFound("No game with this id"),
-                    "privateId" => BadRequest("The given id is not the creator's private one"),
+                    "privateId" => Unauthorized("The given id is not the creator's private one"),
                     _ => BadRequest($"Unexpected error : {ex.Message}")
                 };
             }
